requirements.txt
# Web API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlmodel>=0.0.14
aiomysql>=0.2.0
pymysql>=1.1.0
sqlalchemy[asyncio]>=2.0.0

# Security
cryptography>=41.0.0

# Trading
ccxt>=4.2.0

# Utils
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Distributed Task Queue
celery[redis]>=5.3.0
redis>=5.0.0
